+++ Classes/ViewHelpers/SlotViewHelper.php
@@ -15,8 +15,8 @@ namespace Romm\Formz\ViewHelpers;
 
 use Romm\Formz\Core\Core;
 use Romm\Formz\Exceptions\ContextNotFoundException;
-use Romm\Formz\Service\ViewHelper\FieldViewHelperService;
-use Romm\Formz\Service\ViewHelper\SlotViewHelperService;
+use Romm\Formz\Service\ViewHelper\Field\FieldViewHelperService;
+use Romm\Formz\Service\ViewHelper\Slot\SlotViewHelperService;
 use TYPO3\CMS\Fluid\Core\Rendering\RenderingContextInterface;
 use TYPO3\CMS\Fluid\Core\ViewHelper\Facets\CompilableInterface;
 
@@ -64,7 +64,7 @@ class SlotViewHelper extends AbstractViewHelper implements CompilableInterface
         /** @var SlotViewHelperService $slotService */
         $slotService = Core::instantiate(SlotViewHelperService::class);
 
-        $slotService->addSlot($arguments['name'], $renderChildrenClosure, $arguments['arguments'], $renderingContext);
+        $slotService->addSlot($arguments['name'], $renderChildrenClosure, $arguments['arguments']);
     }
 
     /**